requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.8.1=pyhd8ed1ab_1
black=24.8.0=py313haa95532_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.12.31=haa95532_0
click=8.1.7=py313haa95532_0
colorama=0.4.6=py313haa95532_0
django=5.1.4=pyhd8ed1ab_0
djangorestframework=3.15.2=pyhd8ed1ab_1
flake8=7.1.1=py313haa95532_0
libexpat=2.6.4=he0c23c2_0
libffi=3.4.2=h8ffe710_5
liblzma=5.6.3=h2466b09_1
libmpdec=4.0.0=h2466b09_0
libsqlite=3.47.2=h67fdade_0
libzlib=1.3.1=h2466b09_2
mccabe=0.7.0=pyhd3eb1b0_0
mypy_extensions=1.0.0=py313haa95532_0
openssl=3.4.0=ha4e3fda_1
packaging=24.2=py313haa95532_0
pathspec=0.10.3=py313haa95532_0
pip=24.3.1=pyh145f28c_2
platformdirs=3.10.0=py313haa95532_0
pycodestyle=2.12.1=py313haa95532_0
pyflakes=3.2.0=py313haa95532_0
python=3.13.1=h071d269_104_cp313
python_abi=3.13=5_cp313
sqlparse=0.5.3=pyhd8ed1ab_0
tk=8.6.13=h5226925_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=hc8b5060_0
ucrt=10.0.22621.0=h57928b3_1
vc=14.3=ha32ba9b_23
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
